feat: Add NoSQL language support for VS Code editor and refactor shared definitions
- Extract shared language definitions into src/cosmosdb/language/nosqlLanguageDefinitions.ts (keywords, functions, schema helpers) with no vscode/monaco imports

- Add TextMate grammar (syntaxes/nosql.tmLanguage.json) for VS Code editor syntax highlighting

- Add language-configuration.json for comment toggling, bracket matching, auto-closing pairs

- Add VS Code CompletionItemProvider for keyword and function completions in .nosql files

- Register completion provider at extension activation, add onLanguage:nosql activation event

- Add file icon (CosmosDBAccount.svg) for .nosql files

- Refactor webview Monarch tokenizer and Monaco completion provider to import from shared module

- Add webpack copy rules for syntaxes/ and language-configuration.json

- Replace plan document with docs/nosql-language-support.md describing architecture and file roles
