+++ syntaxes/nosql.tmLanguage.json
@@ -1,0 +1,175 @@
+{
+    "name": "CosmosDB NoSQL",
+    "scopeName": "source.nosql",
+    "patterns": [
+        { "include": "#comments" },
+        { "include": "#strings" },
+        { "include": "#numbers" },
+        { "include": "#operators" },
+        { "include": "#builtinFunctions" },
+        { "include": "#keywords" },
+        { "include": "#identifiers" }
+    ],
+    "repository": {
+        "comments": {
+            "patterns": [
+                {
+                    "name": "comment.line.double-dash.nosql",
+                    "match": "--.*$"
+                },
+                {
+                    "name": "comment.block.nosql",
+                    "begin": "/\\*",
+                    "end": "\\*/"
+                }
+            ]
+        },
+        "strings": {
+            "patterns": [
+                {
+                    "name": "string.quoted.single.nosql",
+                    "begin": "'",
+                    "end": "'",
+                    "patterns": [
+                        {
+                            "name": "constant.character.escape.nosql",
+                            "match": "\\\\."
+                        },
+                        {
+                            "name": "constant.character.escape.nosql",
+                            "match": "''"
+                        }
+                    ]
+                },
+                {
+                    "name": "string.quoted.double.nosql",
+                    "begin": "\"",
+                    "end": "\"",
+                    "patterns": [
+                        {
+                            "name": "constant.character.escape.nosql",
+                            "match": "\\\\."
+                        }
+                    ]
+                }
+            ]
+        },
+        "numbers": {
+            "patterns": [
+                {
+                    "name": "constant.numeric.nosql",
+                    "match": "\\b\\d+(\\.\\d+)?([eE][-+]?\\d+)?\\b"
+                }
+            ]
+        },
+        "operators": {
+            "patterns": [
+                {
+                    "name": "keyword.operator.comparison.nosql",
+                    "match": "!=|<>|<=|>=|[<>=!]"
+                },
+                {
+                    "name": "keyword.operator.arithmetic.nosql",
+                    "match": "[+\\-*/%]"
+                },
+                {
+                    "name": "keyword.operator.logical.nosql",
+                    "match": "\\|\\||&&|\\?\\?"
+                },
+                {
+                    "name": "keyword.operator.ternary.nosql",
+                    "match": "[?:]"
+                }
+            ]
+        },
+        "builtinFunctions": {
+            "patterns": [
+                {
+                    "name": "support.function.aggregate.nosql",
+                    "match": "(?i)\\b(AVG|COUNT|MAX|MIN|SUM)\\b(?=\\s*\\()"
+                },
+                {
+                    "name": "support.function.math.nosql",
+                    "match": "(?i)\\b(ABS|ACOS|ASIN|ATAN|ATN2|CEILING|COS|COT|DEGREES|EXP|FLOOR|LOG|LOG10|PI|POWER|RADIANS|RAND|ROUND|SIGN|SIN|SQRT|SQUARE|TAN|TRUNC)\\b(?=\\s*\\()"
+                },
+                {
+                    "name": "support.function.math.integer.nosql",
+                    "match": "\\b(IntAdd|IntBitAnd|IntBitLeftShift|IntBitNot|IntBitOr|IntBitRightShift|IntBitXor|IntDiv|IntMod|IntMul|IntSub|NumberBin)\\b(?=\\s*\\()"
+                },
+                {
+                    "name": "support.function.typecheck.nosql",
+                    "match": "(?i)\\b(IS_ARRAY|IS_BOOL|IS_DEFINED|IS_FINITE_NUMBER|IS_INTEGER|IS_NULL|IS_NUMBER|IS_OBJECT|IS_PRIMITIVE|IS_STRING)\\b(?=\\s*\\()"
+                },
+                {
+                    "name": "support.function.string.nosql",
+                    "match": "(?i)\\b(CONCAT|CONTAINS|ENDSWITH|INDEX_OF|LEFT|LENGTH|LOWER|LTRIM|REGEXMATCH|REPLACE|REPLICATE|REVERSE|RIGHT|RTRIM|STARTSWITH|SUBSTRING|TRIM|UPPER)\\b(?=\\s*\\()"
+                },
+                {
+                    "name": "support.function.string.nosql",
+                    "match": "\\b(StringEquals|ToString)\\b(?=\\s*\\()"
+                },
+                {
+                    "name": "support.function.array.nosql",
+                    "match": "(?i)\\b(ARRAY_CONCAT|ARRAY_CONTAINS|ARRAY_LENGTH|ARRAY_SLICE)\\b(?=\\s*\\()"
+                },
+                {
+                    "name": "support.function.array.nosql",
+                    "match": "\\b(SetIntersect|SetUnion)\\b(?=\\s*\\()"
+                },
+                {
+                    "name": "support.function.datetime.nosql",
+                    "match": "\\b(DateTimeAdd|DateTimeBin|DateTimeDiff|DateTimeFromParts|DateTimePart|DateTimeToTicks|DateTimeToTimestamp|GetCurrentDateTime|GetCurrentDateTimeStatic|GetCurrentTicks|GetCurrentTicksStatic|GetCurrentTimestamp|GetCurrentTimestampStatic|TicksToDateTime)\\b(?=\\s*\\()"
+                },
+                {
+                    "name": "support.function.spatial.nosql",
+                    "match": "(?i)\\b(ST_AREA|ST_DISTANCE|ST_INTERSECTS|ST_ISVALID|ST_ISVALIDDETAILED|ST_WITHIN)\\b(?=\\s*\\()"
+                },
+                {
+                    "name": "support.function.object.nosql",
+                    "match": "\\b(ObjectToArray|AllMembers)\\b(?=\\s*\\()"
+                },
+                {
+                    "name": "support.function.conversion.nosql",
+                    "match": "\\b(StringToArray|StringToBoolean|StringToNull|StringToNumber|StringToObject)\\b(?=\\s*\\()"
+                },
+                {
+                    "name": "support.function.vector.nosql",
+                    "match": "\\b(VectorDistance)\\b(?=\\s*\\()"
+                }
+            ]
+        },
+        "keywords": {
+            "patterns": [
+                {
+                    "name": "keyword.clause.nosql",
+                    "match": "(?i)\\b(SELECT|DISTINCT|TOP|FROM|WHERE|ORDER\\s+BY|GROUP\\s+BY|HAVING|OFFSET|LIMIT|JOIN)\\b"
+                },
+                {
+                    "name": "keyword.other.nosql",
+                    "match": "(?i)\\b(IN|AS|VALUE|EXISTS|BETWEEN|LIKE|ARRAY|UDF)\\b"
+                },
+                {
+                    "name": "keyword.operator.logical.nosql",
+                    "match": "(?i)\\b(AND|OR|NOT)\\b"
+                },
+                {
+                    "name": "keyword.other.order.nosql",
+                    "match": "(?i)\\b(ASC|DESC)\\b"
+                },
+                {
+                    "name": "constant.language.nosql",
+                    "match": "(?i)\\b(NULL|TRUE|FALSE|UNDEFINED)\\b"
+                }
+            ]
+        },
+        "identifiers": {
+            "patterns": [
+                {
+                    "name": "variable.other.nosql",
+                    "match": "\\b[a-zA-Z_]\\w*\\b"
+                }
+            ]
+        }
+    }
+}
+
